Move map tree updating code to a separate file
